Move Zone::YamlDefinitions to its own file
