+++ lib/record_store.rb
@@ -21,6 +21,7 @@ require 'record_store/record/ns'
 require 'record_store/record/txt'
 require 'record_store/record/spf'
 require 'record_store/record/srv'
+require 'record_store/zone/yaml_definitions'
 require 'record_store/zone'
 require 'record_store/zone/config'
 require 'record_store/changeset'
